Reorganization of test projects.
