+++ test/FunctionalTests/Diagnostics.HelthChecks/HealthCheckApplicationBuilderExtensionsTests.cs
@@ -54,47 +54,7 @@ namespace Diagnostics.HelthChecks
                 content.Should().Be("Healthy");
             }
         }
-
-        [Fact]
-        public void HealthCheckApplicationBuilderExtensionsWithAppNull()
-        {
-            FluentActions.Invoking(() =>
-                HealthCheckApplicationBuilderExtensions.UseHealthChecks(null, null, new PathString(""))
-            )
-                .Should()
-                .Throw<ArgumentNullException>();
-        }
-
-        [Fact]
-        public void HealthCheckApplicationBuilderExtensionsWithAppAndOptionsNull()
-        {
-            FluentActions.Invoking(() =>
-                HealthCheckApplicationBuilderExtensions.UseHealthChecks(null, null, new PathString(""), options: null)
-            )
-                .Should()
-                .Throw<ArgumentNullException>();
-        }
-
-        [Fact]
-        public void HealthCheckApplicationBuilderExtensionsWithAppNullAndPort()
-        {
-            FluentActions.Invoking(() =>
-                HealthCheckApplicationBuilderExtensions.UseHealthChecks(null, null, new PathString(""), 25)
-            )
-                .Should()
-                .Throw<ArgumentNullException>();
-        }
-
-        [Fact]
-        public void HealthCheckApplicationBuilderExtensionsWithAppAndOptionsNullAndPort()
-        {
-            FluentActions.Invoking(() =>
-                HealthCheckApplicationBuilderExtensions.UseHealthChecks(null, null, new PathString(""), 25, null)
-            )
-                .Should()
-                .Throw<ArgumentNullException>();
-        }
-
+      
         #region [ Arrange ]
 
         public class StartupWithhHealthCheckOptions : Startup
