+++ test/UnitTests/Diagnostics.HelthChecks/Diagnostics.HelthChecks.UnitTests.csproj
@@ -23,8 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Diagnostics.HealthChecks\Diagnostics.HealthChecks.csproj" />
-    <ProjectReference Include="..\Common\Common.csproj" />
+    <ProjectReference Include="..\..\..\src\Diagnostics.HealthChecks\Diagnostics.HealthChecks.csproj" />
   </ItemGroup>
 
 </Project>
